feat: Add pagination component to restaurant listings
This commit adds a new pagination component to the restaurant listings page. The pagination component will be used to navigate through the list of restaurants. It is implemented as a separate component called "Pagination" and is imported in the "restaurant-listings.tsx" file. This addition will enhance the user experience by allowing them to easily navigate through multiple pages of restaurant listings.